Optimization of SELF file listing. Only .bin, .elf, .sprx, .prx and .self files are checked.

+++ self.h
@@ -99,7 +99,7 @@ int self_extract_elf_ex(int self_fd, int elf_fd, int verify);
 /**
  * Check if the given path is a SELF file.
  **/
-int self_is_valid(const char* path);
+size_t self_is_valid(const char* path);
 
 
 /**
